Fix RGB decay stalls and master-side freeze in Pastel Pulse effect
- Fix decay stall bug by using uint32_t for ticks and decay calculations in matrix_scan_user, preventing overflow during long idle periods.
- Fix master-side freeze by updating the effect return value to handle empty rendering chunks on split master halves.
- Add intermediate uint64_t casting to animation time math to prevent multiplication overflows with speed.
- Maintain critical bounds checking and volatile usage for stability.

+++ keyboards/sofle_choc/keymaps/vial/keymap.c
@@ -161,13 +161,13 @@ void matrix_scan_user(void) {
     // Robust decay: every 10ms
     uint32_t elapsed = timer_elapsed32(decay_timer);
     if (elapsed >= 10) {
-        uint8_t ticks = elapsed / 10;
+        uint32_t ticks = elapsed / 10;
         decay_timer += ticks * 10;
         for (uint8_t i = 0; i < RGB_MATRIX_LED_COUNT; i++) {
             if (led_boost[i] > 127) {
-                uint8_t decay = ticks * 4;
+                uint32_t decay = ticks * 4;
                 if (led_boost[i] > 127 + decay) {
-                    led_boost[i] -= decay;
+                    led_boost[i] -= (uint8_t)decay;
                 } else {
                     led_boost[i] = 127;
                 }
